requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.88.0
h11==0.14.0
idna==3.4
importlib-metadata==5.1.0
ln-markets==1.0.13
peewee==3.15.4
pydantic==1.10.2
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.1
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
websocket-client==1.4.2
zipp==3.11.0
